bugfix0.09

+++ ais-bench_workload/tool/find_best_batchsize/test/test_aoe_mode.sh
@@ -20,8 +20,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[[ $cmd != $ret_ok ]] && { echo "command failed"; return 1 }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -37,9 +36,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 0
     --job_type 1
     "
-$cmd
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[[ $cmd != $ret_ok ]] && { echo "command failed"; return 1 }
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -56,8 +53,8 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[[ $cmd != $ret_ok ]] && { echo "command failed"; return 1 }
+
 cmd="rm -rf ../cache/"
 $cmd
 
@@ -75,8 +72,7 @@ cmd="bash ../find_best_batchsize.sh
     --aoe_mode 1
     --job_type 1
     "
-# [[ $cmd != ret_ok ]] && { echo "command failed"; return 1 }
-$cmd
+[[ $cmd != $ret_ok ]] && { echo "command failed"; return 1 }
 cmd="rm -rf ../cache/"
 $cmd
